requirement.txt
streamlit
pandas
pyyaml
fpdf
openai
python-docx

